Move object index handling to objectmgr.
